+++ check.sh
@@ -1,3 +1,0 @@
-#!/bin/bash
-cd /Users/joseph/GIT/nzi-cli
-cargo check 2>&1